+++ src/matrix/matrix.simd.go
@@ -1,4 +1,4 @@
-//go:build (amd64 || arm64) && !F64
+//go:build amd64 || arm64
 
 /******************************************************************************/
 /* matrix.simd.go                                                             */
